Handle HTTP errors properly
I've got a number of strange `checksum mismatch, re-downloading` errors
on my repository and decided to investigate the reason. It turned out
that current implementation of googlephotos module doesn't check HTTP
status code of HTTP response and writes body of non-`200 OK` HTTP
response in place of the photo to repository (!) + there is a bug
in `repo.go` due to which all download errors are ignored because
invalid `err` is used.